fix: Handle proper type conversion on sort nodes. (#228)
* Handled CBOR encoding edge case

* Fixed tests which manually casted results to uint64 due to previous encoding inconsistency

+++ db/tests/query/one_to_two_many/simple_test.go
@@ -309,7 +309,7 @@ func TestQueryOneToTwoManyWithNamedAndUnnamedRelationships(t *testing.T) {
 					},
 					"price": map[string]interface{}{
 						"currency": "SEK",
-						"value":    uint64(129),
+						"value":    float64(129),
 					},
 				},
 				{
@@ -414,7 +414,7 @@ func TestQueryOneToTwoManyWithNamedAndUnnamedRelationships(t *testing.T) {
 						{
 							"name": "A Time for Mercy",
 							"price": map[string]interface{}{
-								"value": uint64(129),
+								"value": float64(129),
 							},
 						},
 						{
